refactor(interactive_shell): split command_registry/rca_cmds.py into rca/ package (#5446)
Split the monolithic surfaces/interactive_shell/command_registry/rca_cmds.py module (489 lines) into a structured surfaces/interactive_shell/command_registry/rca/ package following the pattern in command_registry/model/ and command_registry/agents/.

- rca/__init__.py: Package exports (COMMANDS) under 30 lines (7 lines).
- rca/records.py: Finding, resolving, and formatting stored RCA record metadata and headers.
- rca/export.py: Path normalization and file export for RCA investigation reports.
- rca/menu.py: Interactive TTY menus and pickers for persisted RCA reports.
- rca/command.py: Slash command handlers and registration for /rca.
- Remove deprecated rca_cmds.py file and update canonical import paths across command_registry/__init__.py, help.py, and test_rca_cmds.py.

Fixes #5446.
