+++ surfaces/interactive_shell/command_registry/rca/__init__.py
@@ -1,0 +1,7 @@
+"""Slash command: ``/rca`` (browse persisted RCA reports across sessions)."""
+
+from __future__ import annotations
+
+from surfaces.interactive_shell.command_registry.rca.command import COMMANDS
+
+__all__ = ["COMMANDS"]
