+++ tests/interactive_shell/test_rca_cmds.py
@@ -79,7 +79,8 @@ def test_tty_rca_menu_latest_shows_report(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    from surfaces.interactive_shell.command_registry import rca_cmds
+    from surfaces.interactive_shell.command_registry.rca import command as rca_command
+    from surfaces.interactive_shell.command_registry.rca import menu as rca_menu
 
     monkeypatch.setattr("config.constants.OPENSRE_HOME_DIR", tmp_path)
     monkeypatch.setattr("config.constants.paths.OPENSRE_HOME_DIR", tmp_path)
@@ -91,8 +92,8 @@ def test_tty_rca_menu_latest_shows_report(
         trigger="/investigate generic",
     )
 
-    monkeypatch.setattr(rca_cmds, "repl_tty_interactive", lambda: True)
-    monkeypatch.setattr(rca_cmds, "repl_choose_one", lambda **_: rca_cmds._RCA_LATEST)
+    monkeypatch.setattr(rca_command, "repl_tty_interactive", lambda: True)
+    monkeypatch.setattr(rca_menu, "repl_choose_one", lambda **_: rca_menu._RCA_LATEST)
 
     console, buf = _capture()
     assert dispatch_slash("/rca", session, console) is True
@@ -106,7 +107,8 @@ def test_tty_rca_history_menu_picks_report_directly(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    from surfaces.interactive_shell.command_registry import rca_cmds
+    from surfaces.interactive_shell.command_registry.rca import command as rca_command
+    from surfaces.interactive_shell.command_registry.rca import menu as rca_menu
 
     monkeypatch.setattr("config.constants.OPENSRE_HOME_DIR", tmp_path)
     monkeypatch.setattr("config.constants.paths.OPENSRE_HOME_DIR", tmp_path)
@@ -123,8 +125,8 @@ def test_tty_rca_history_menu_picks_report_directly(
         trigger="/investigate generic",
     )
 
-    monkeypatch.setattr(rca_cmds, "repl_tty_interactive", lambda: True)
-    monkeypatch.setattr(rca_cmds, "repl_choose_one", lambda **_: older_id)
+    monkeypatch.setattr(rca_command, "repl_tty_interactive", lambda: True)
+    monkeypatch.setattr(rca_menu, "repl_choose_one", lambda **_: older_id)
 
     console, buf = _capture()
     assert dispatch_slash("/rca history", session, console) is True
@@ -138,7 +140,8 @@ def test_tty_rca_root_menu_history_picks_report(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    from surfaces.interactive_shell.command_registry import rca_cmds
+    from surfaces.interactive_shell.command_registry.rca import command as rca_command
+    from surfaces.interactive_shell.command_registry.rca import menu as rca_menu
 
     monkeypatch.setattr("config.constants.OPENSRE_HOME_DIR", tmp_path)
     monkeypatch.setattr("config.constants.paths.OPENSRE_HOME_DIR", tmp_path)
@@ -155,9 +158,9 @@ def test_tty_rca_root_menu_history_picks_report(
         trigger="/investigate generic",
     )
 
-    picks = iter([rca_cmds._RCA_HISTORY, older_id])
-    monkeypatch.setattr(rca_cmds, "repl_tty_interactive", lambda: True)
-    monkeypatch.setattr(rca_cmds, "repl_choose_one", lambda **_: next(picks))
+    picks = iter([rca_menu._RCA_HISTORY, older_id])
+    monkeypatch.setattr(rca_command, "repl_tty_interactive", lambda: True)
+    monkeypatch.setattr(rca_menu, "repl_choose_one", lambda **_: next(picks))
 
     console, buf = _capture()
     assert dispatch_slash("/rca", session, console) is True
@@ -307,9 +310,9 @@ def test_rca_save_unknown_id_reports_not_found(
 
 
 def test_normalize_rca_save_path_strips_quotes() -> None:
-    from surfaces.interactive_shell.command_registry import rca_cmds
+    from surfaces.interactive_shell.command_registry.rca import export as rca_export
 
-    dest = rca_cmds._normalize_rca_save_path("'/tmp/report.md'", investigation_id="abcd1234")
+    dest = rca_export._normalize_rca_save_path("'/tmp/report.md'", investigation_id="abcd1234")
     assert dest == Path("/tmp/report.md")
 
 
@@ -317,7 +320,8 @@ def test_tty_rca_save_menu_picks_latest_and_prompts_path(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    from surfaces.interactive_shell.command_registry import rca_cmds
+    from surfaces.interactive_shell.command_registry.rca import command as rca_command
+    from surfaces.interactive_shell.command_registry.rca import menu as rca_menu
 
     monkeypatch.setattr("config.constants.OPENSRE_HOME_DIR", tmp_path)
     monkeypatch.setattr("config.constants.paths.OPENSRE_HOME_DIR", tmp_path)
@@ -330,9 +334,9 @@ def test_tty_rca_save_menu_picks_latest_and_prompts_path(
     )
 
     dest = tmp_path / "picked.md"
-    monkeypatch.setattr(rca_cmds, "repl_tty_interactive", lambda: True)
-    monkeypatch.setattr(rca_cmds, "repl_choose_one", lambda **_: rca_cmds._RCA_LATEST)
-    monkeypatch.setattr(rca_cmds, "_prompt_rca_save_path", lambda _console: str(dest))
+    monkeypatch.setattr(rca_command, "repl_tty_interactive", lambda: True)
+    monkeypatch.setattr(rca_menu, "repl_choose_one", lambda **_: rca_menu._RCA_LATEST)
+    monkeypatch.setattr(rca_menu, "_prompt_rca_save_path", lambda _console: str(dest))
 
     console, buf = _capture()
     assert dispatch_slash("/rca save", session, console) is True
